services/requirements.txt
# === Core API ===
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sse-starlette>=2.0.0
pydantic[email]>=2.6.0
python-multipart>=0.0.9

# === Auth ===
PyJWT>=2.8.0
bcrypt>=4.1.0

# === Database ===
sqlalchemy[asyncio]>=2.0.28
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# === Redis ===
redis>=5.0.0

# === Object Storage ===
boto3>=1.34.0

# === HTTP client (worker -> inference) ===
httpx>=0.27.0

# === ML / ASR (inference service) ===
faster-whisper>=1.0.0

# === Video processing (worker) ===
ffmpeg-python>=0.2.0
yt-dlp>=2025.1.15

# === Existing project deps ===
numpy>=1.26.0
scipy>=1.12.0
rich>=13.7.0
